Added some basic RHEL/CentOS 7 32bit libraries
E.g. /app/arc/2.3/bin/addproj links in the end to glibc.i686.
(This is ARC, not AFS specific.)
This could happend on any RHEL/CentOS 7 system, even though it only
now showed for the first time on global CentOS 7 profile. We don't
expect a collision with other modules for that package.